requirements.txt
catboost==1.2.8
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.2
graphviz==0.20.3
imbalanced-learn==0.13.0
joblib==1.5.1
kiwisolver==1.4.8
lightgbm==4.6.0
matplotlib==3.10.3
narwhals==1.41.1
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
tabulate==0.9.0
threadpoolctl==3.6.0
tzdata==2025.2
xgboost==3.0.2
